backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
whitenoise>=6.6
gunicorn>=22.0
httpx>=0.27
openpyxl>=3.1
python-docx>=1.1
pymupdf>=1.24
dj-database-url>=2.1
psycopg2-binary>=2.9
